+++ src/lib/logic/ports/auth-services-interfaces.ts
@@ -1,5 +1,10 @@
 import type { Role, User } from '$lib/entities/user';
 
+export interface Hash_Service {
+	hash(data: string): Promise<string>;
+	check(x: string, y: string): Promise<{ pass: boolean }>;
+}
+
 export type Payload = {
 	id: number;
 	username: string;
