--- src/lib/logic/ports/i-basic-repo.ts
@@ -1,6 +1,0 @@
-export interface Basic_Repo<T extends { id: number }> {
-	get_all(): Promise<T[]>;
-	get_one(id: number): Promise<T | undefined>;
-	remove(id: number): Promise<T | undefined>;
-	update(entity: T): Promise<T | undefined>;
-}
