--- src/lib/logic/ports/i-hash-service.ts
@@ -1,4 +1,0 @@
-export interface Hash_Service {
-	hash(data: string): Promise<string>;
-	check(x: string, y: string): Promise<{ pass: boolean }>;
-}
