+++ src/lib/logic/ports/repos-interfaces.ts
@@ -1,0 +1,29 @@
+import type { Category } from '$lib/entities/category';
+import type { User } from '$lib/entities/user';
+import type { Role } from '$lib/entities/user';
+
+export type Login_Response = { pass: true; user: User } | { pass: false };
+
+export interface User_Repo {
+	get_one(id: number): Promise<User | undefined>;
+	get_all(): Promise<User[]>;
+	create(user: { username: string; password: string }): Promise<User>;
+	validate(user: { username: string; password: string }): Promise<Login_Response>;
+	add_role(data: { user_id: number; role_id: number }): Promise<void>;
+	remove_all_roles(data: { user_id: number }): Promise<void>;
+	get_by_username(username: string): Promise<User | undefined>;
+}
+
+export interface Role_Repo {
+	get_all(): Promise<Role[]>;
+	get_one(id: number): Promise<Role | undefined>;
+	create(category: { id: number; name: string }): Promise<Role>;
+}
+
+export interface Category_Repo {
+	get_all(): Promise<Category[]>;
+	get_one(id: number): Promise<Category | undefined>;
+	create(category: { name: string }): Promise<Category>;
+	get_by_name(name: string): Promise<Category | undefined>;
+	update(modified: Category): Promise<void>;
+}
