+++ src/lib/repos/role-repo-drizzle.ts
@@ -1,6 +1,6 @@
 import { eq } from 'drizzle-orm';
 import type { Role } from '$lib/entities/user';
-import type { Role_Repo } from '$lib/logic/ports/i-role-repo';
+import type { Role_Repo } from '$lib/logic/ports/repos-interfaces';
 import type { DB_Context } from '$lib/server/drizzle/drizzle-client';
 import { t_role } from '$lib/server/drizzle/schema';
