+++ src/lib/repos/unit-of-work.ts
@@ -1,6 +1,6 @@
 import { TransactionDatabaseError } from '$lib/errors';
-import type { Hash_Service } from '$lib/logic/ports/i-hash-service';
-import type { Repos, Unit_of_Work } from '$lib/logic/ports/i-unit-of-work';
+import type { Hash_Service } from '$lib/logic/ports/auth-services-interfaces';
+import type { Repos, Unit_of_Work } from '$lib/logic/ports/unit-of-work-interface';
 import type { DB_Context } from '$lib/server/drizzle/drizzle-client';
 import { Category_Repo_Drizzle } from './category-repo-drizzle';
 import { User_Repo_Drizzle } from './user-repo-drizzle';
