+++ src/lib/services/jwt_service.ts
@@ -1,7 +1,11 @@
 import { JWT_SECRET_KEY } from '$env/static/private';
 import { SignJWT, jwtVerify } from 'jose';
 import { User } from '$lib/entities/user';
-import type { Payload, Token_Service, Token_Validate_Res } from '$lib/logic/ports/i-token-service';
+import type {
+	Payload,
+	Token_Service,
+	Token_Validate_Res
+} from '$lib/logic/ports/auth-services-interfaces';
 
 export class JWT_Service implements Token_Service {
 	async validate(token: string): Promise<Token_Validate_Res> {
